… proposals/optimization-by-compiling-regex-strategy.md
… proposals/optimizations-from-rust-regex.md
… src/regex/prefilter.mojo