This PR implements a prefilter system inspired by the Rust `regex-automata` crate to improve regex performance through fast candidate identification before full pattern matching.

## Summary

### Key Components

#### **Prefilter System** (`src/regex/prefilter.mojo`)
- **LiteralExtractor**: Extracts required literals, prefixes, and suffixes from regex patterns
- **MemchrPrefilter**: Fast literal scanning using optimized string search
- **PrefilterMatcher**: Coordinates prefilter application with regex engines

#### **Enhanced Matcher** (`src/regex/matcher.mojo`) 
- **Prefilter integration**: Seamlessly integrates literal prefiltering with existing engines
- **Performance heuristics**: Skips prefilter analysis for simple patterns where overhead > benefit
- **Exact literal bypass**: Direct string matching for pure literal patterns

#### **Architecture Documentation** (`proposals/`)
- **Deep analysis**: Comparison with Rust regex-automata design patterns
- **Implementation strategy**: Phased approach for systematic optimization

## Technical Architecture

### Prefilter Flow
```mojo
# Fast candidate identification before full regex matching
struct PrefilterMatcher:
    fn find_candidates(self, text: String) -> List[Int]
    fn apply_prefilter(self, text: String, engine: RegexMatcher) -> List[Match]
```

### Engine Coordination
```mojo  
fn create_optimized_matcher(pattern: String) -> RegexMatcher:
    var prefilter = create_prefilter(pattern)
    if prefilter.is_effective():
        return PrefilterMatcher(prefilter, select_best_engine(pattern))
    return select_best_engine(pattern)
```

### Performance results

```
--------------------------------------------------------------------------------------------------------------                                                                                                                                                                                                                
Benchmark                           Other Branch (ms)     Branch (ms)    Speedup   Engine          Status                                                                                                                                                                                                                     
--------------------------------------------------------------------------------------------------------------                                                                                                                                                                                                                
alternation_common_prefix                  0.117 ms        0.119 ms      0.98x      DFA       ≈ Similar                                                                                                                                                                                                                       
alternation_simple                         0.204 ms        0.210 ms      0.97x      DFA       ≈ Similar                                                                                                                                                                                                                       
alternation_words                          0.267 ms        0.316 ms      0.85x      DFA ← Other Branch slight                                                                                                                                                                                                                 
anchor_end                                 0.443 ms        0.467 ms      0.95x      DFA       ≈ Similar                                                                                                                                                                                                                       
anchor_start                               0.479 ms        0.472 ms      1.01x      DFA       ≈ Similar                                                                                                                                                                                                                       
complex_email_extraction                  237.81 ms       249.33 ms      0.95x      DFA       ≈ Similar                                                                                                                                                                                                                       
complex_number_extraction                 209.93 ms       238.19 ms      0.88x      DFA ← Other Branch slight                                                                                                                                                                                                                 
group_alternation                          0.228 ms        0.222 ms      1.03x      DFA       ≈ Similar                                                                                                                                                                                                                       
group_quantified                           0.236 ms        0.245 ms      0.97x      DFA       ≈ Similar                                                                                                                                                                                                                       
literal_match_long                         0.444 ms        0.565 ms      0.79x      DFA ← Other Branch slight                                                                                                                                                                                                                 
literal_match_short                        0.439 ms        0.459 ms      0.96x      DFA       ≈ Similar                                                                                                                                                                                                                       
literal_prefix_long                        0.869 ms        0.938 ms      0.93x      NFA       ≈ Similar                                                                                                                                                                                                                       
literal_prefix_medium                      0.089 ms        0.093 ms      0.95x      NFA       ≈ Similar                                                                                                                                                                                                                       
literal_prefix_short                       0.029 ms        0.341 μs     85.52x NFA+ExactLiteral  🚀 Branch wins!                                                                                                                                                                                                              
match_all_pattern                          11.15 ms        11.43 ms      0.98x      DFA       ≈ Similar                                                                                                                                                                                                                       
match_all_simple                           10.36 ms        10.26 ms      1.01x      DFA       ≈ Similar                                                                                                                                                                                                                       
no_literal_baseline                        0.108 ms        0.104 ms      1.04x      DFA       ≈ Similar                                                                                                                                                                                                                       
quantifier_one_or_more                     0.219 ms        0.226 ms      0.97x      DFA       ≈ Similar                                                                                                                                                                                                                       
quantifier_zero_or_more                    0.209 ms        0.233 ms      0.90x      DFA ← Other Branch slight                                                                                                                                                                                                                 
quantifier_zero_or_one                     0.219 ms        0.249 ms      0.88x      DFA ← Other Branch slight                                                                                                                                                                                                                 
range_alphanumeric                         54.86 ms        57.60 ms      0.95x      DFA       ≈ Similar                                                                                                                                                                                                                       
range_digits                               0.226 ms        0.233 ms      0.97x      DFA       ≈ Similar                                                                                                                                                                                                                       
range_lowercase                            0.308 ms        0.258 ms      1.19x      DFA → Branch slight                                                                                                                                                                                                                       
required_literal_long                      0.009 ms        0.010 ms      0.93x NFA+Prefilter       ≈ Similar                                                                                                                                                                                                                  
required_literal_short                     0.443 μs        0.534 μs      0.83x NFA+Prefilter ← Other Branch slight                                                                                                                                                                                                            
simd_alphanumeric_large                    0.058 ms        0.055 ms      1.06x      DFA       ≈ Similar                                                                                                                                                                                                                       
simd_alphanumeric_xlarge                   0.054 ms        0.059 ms      0.91x      DFA       ≈ Similar                                                                                                                                                                                                                       
simd_multi_char_class                      0.038 ms        0.046 ms      0.82x      DFA ← Other Branch slight                                                                                                                                                                                                                 
simd_negated_alphanumeric                  0.040 ms        0.044 ms      0.90x      DFA       ≈ Similar                                                                                                                                                                                                                       
wildcard_match_any                         61.96 ms        0.137 ms    452.12x      NFA  🚀 Branch wins!                                                                                                                                                                                                                      
--------------------------------------------------------------------------------------------------------------                                                                                                                                                                                                                
                                                                                                                                                                                                                                                                                                                              
TOP 5 SPEEDUPS (Branch vs Other Branch):                                                                                                                                                                                                                                                                                      
  1. wildcard_match_any: 452.12x faster                                                                                                                                                                                                                                                                                       
  2. literal_prefix_short: 85.52x faster                                                                                                                                                                                                                                                                                      
  3. range_lowercase: 1.19x faster                                                                                                                                                                                                                                                                                            
  4. simd_alphanumeric_large: 1.06x faster                                                                                                                                                                                                                                                                                    
  5. no_literal_baseline: 1.04x faster                                                                                                                                                                                                                                                                                        
                                                                                                                                                                                                                                                                                                                              
BOTTOM 5 SPEEDUPS:                                                                                                                                                                                                                                                                                                            
  1. quantifier_zero_or_one: 0.88x                                                                                                                                                                                                                                                                                            
  2. alternation_words: 0.85x                                                                                                                                                                                                                                                                                                 
  3. required_literal_short: 0.83x                                                                                                                                                                                                                                                                                            
  4. simd_multi_char_class: 0.82x                                                                                                                                                                                                                                                                                             
  5. literal_match_long: 0.79x   
```